user-service/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-Cors==4.0.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
bcrypt==4.0.1
PyJWT==2.8.0
pytz==2025.1